make file bien e  itoa ^^

+++ Libft/libft.a/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: mnieto-m <mnieto-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/25 19:57:54 by mnieto-m          #+#    #+#             */
-/*   Updated: 2024/02/26 00:25:20 by mnieto-m         ###   ########.fr       */
+/*   Updated: 2024/02/27 16:37:01 by mnieto-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dst, char *src, size_t dstsize)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (dstsize > 0)
